requirements.txt
crewai
crewai_tools
load_dotenv
langchain
langchain-huggingface
langchain-openai